Export non-default ERC20 ABI.
Default exports are difficult to discover.  Instead, export well named items from each library, keeping in mind that users of the library will be doing `import { name } from 'package'`.

In this case, naming the variable `ERC20ABI` provides additional clarity since it will be common to import several ABIs.  If the user wants to change the name, they can do `import { ERC20ABI as whatever } from 'library'`.

Note: I personally prefer `Erc20Abi`, but it appears the style in this repository is all caps for acronyms so I have followed that in this PR.